Add files via upload
Programs loaded on startup or from LOAD command have a re-scan of local labels.  This is necessary because if a program is run before the re-scan is done, it will not work.  Upon entry in program edit mode, and subsequent exit, an automatic re-scan is done.  This fix eliminates the need to enter and exit program edit mode.